refactor: Update Dockerfile to include go get command for gqlgen
This commit updates the Dockerfile to include a `go get` command for the gqlgen package. Previously, only the `go run` command was used to generate the graphql schema. By adding the `go get` command, we ensure that the gqlgen package is installed before running the generate command. This helps to avoid any potential issues with missing dependencies during the build process.